In Score changed return type from TreeSet to List. Test to mySql was Added.

+++ src/MBean/AccountServiceController.java
@@ -22,7 +22,7 @@ public class AccountServiceController implements AccountServiceControllerMBean {
     }
 
     @Override
-    public int getUsers() {
+    public long getUsers() {
         return accountServer.getAmountOfUsers();
     }
